feat: unify render and generate into a single generate command (#50)
Consolidate render and generate into a unified `generate` command with
subcommands: templates, guardrails, interface. Running `generate`
without arguments executes all targets. The top-level `render` command
is preserved as a deprecated alias for `generate templates`.
